requirements.txt
absl-py==1.0.0+computecanada
cachetools==4.2.4+computecanada
certifi==2021.10.8+computecanada
charset-normalizer==2.0.12+computecanada
google-auth==2.3.3+computecanada
google-auth-oauthlib==0.4.6+computecanada
grpcio==1.38.1+computecanada
idna==3.3+computecanada
importlib-metadata==4.11.3+computecanada
markdown==3.3.6+computecanada
nibabel==3.2.1+computecanada
numpy==1.21.4+computecanada
oauthlib==3.1.1+computecanada
opencv-python==4.5.1.48+computecanada
packaging==21.3+computecanada
Pillow-SIMD==7.0.0.post3+computecanada
protobuf==3.19.4+computecanada
pyasn1==0.4.8+computecanada
pyasn1-modules==0.2.8+computecanada
pyparsing==3.0.8+computecanada
requests==2.27.1+computecanada
requests-oauthlib==1.3.0+computecanada
rsa==4.8+computecanada
scipy==1.7.3+computecanada
six==1.16.0+computecanada
tensorboard==2.8.0+computecanada
tensorboard-data-server==0.6.1+computecanada
tensorboard-plugin-wit==1.8.0+computecanada
torch==1.10.0+computecanada
torchvision==0.11.1+computecanada
typing-extensions==4.2.0+computecanada
urllib3==1.26.9+computecanada
werkzeug==2.1.2+computecanada
zipp==3.8.0+computecanada
